+++ utils/__init__.py
@@ -10,3 +10,4 @@ from .llm_api import *
 from .log import *
 from .prompt_override import *
 from .subtitles import *
+from .multitrack import *